1.0.2.0
- Added Track corner names for Red Bull Ring.
- Added Setup Converters for all 6 new GT2 cars(for viewing and comparing setups).
- Added Steering Locks for all 6 new GT2 cars.

+++ Race_Element.Data.ACC/Config/CameraSettingService.cs
@@ -65,7 +65,7 @@ public class CameraSettingService
         public double DashcamFactor { get; set; }
 
         [JsonProperty("panniniProjection")]
-        public int PanniniProjection { get; set; }
+        public float PanniniProjection { get; set; }
 
         [JsonProperty("helicamDistance")]
         public int HelicamDistance { get; set; }
